sim/host: move host implement to posix directory
